refactor(shared-data, robot-server): relocate data file types to shared data
