+++ robot-server/tests/data_files/test_file_auto_deleter.py
@@ -6,7 +6,7 @@ from decoy import Decoy
 
 from robot_server.data_files.data_files_store import DataFilesStore
 from robot_server.data_files.file_auto_deleter import DataFileAutoDeleter
-from robot_server.data_files.models import DataFileSource
+from opentrons_shared_data.data_files import DataFileSource
 from robot_server.deletion_planner import DataFileDeletionPlanner, FileUsageInfo
