+++ robot-server/tests/protocols/test_completed_analysis_store.py
@@ -8,7 +8,7 @@ import pytest
 from sqlalchemy.engine import Engine
 from decoy import Decoy
 
-from robot_server.data_files.models import DataFileSource
+from opentrons_shared_data.data_files import DataFileSource, DataFileInfo
 from robot_server.persistence.tables import (
     analysis_table,
     analysis_primitive_type_rtp_table,
@@ -24,7 +24,6 @@ from opentrons.protocol_reader import (
 )
 from robot_server.data_files.data_files_store import (
     DataFilesStore,
-    DataFileInfo,
 )
 from robot_server.protocols.analysis_memcache import MemoryCache
 from robot_server.protocols.analysis_models import (
